+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Next
 - Fixed a rounding issue causing the world boundary to be offset by 1 block
+- Fixed two sides of the world not rendering
 
 ## 0.0.20
 *2024-08-08 19:52*